move mdp files
