--- exts/aiffelator/aiffelator/tasks/manipulation/picknplace/mdp/__init__.py
@@ -1,16 +1,0 @@
-# Copyright (c) 2022-2024, The Isaac Lab Project Developers.
-# All rights reserved.
-#
-# SPDX-License-Identifier: BSD-3-Clause
-
-"""This sub-module contains the functions that are specific to the lift environments."""
-
-from omni.isaac.lab.envs.mdp import *  # noqa: F401, F403
-
-from .actions import *  # noqa: F401, F403
-from .commands import *  # noqa: F401, F403
-from .events import *  # noqa: F401, F403
-from .observations import *  # noqa: F401, F403
-from .rewards import *  # noqa: F401, F403
-from .scenes import *  # noqa: F401, F403
-from .terminations import *  # noqa: F401, F403
